Refactor extension tests to run with staging deployment
- Remove extension integration tests
- Integrate extension loading tests into staging deployment tests
- Make tests more realistic and similar to component tests
- Add ExtensionPage class for better test organization
- Update GitHub Actions workflow and package.json
