+++ .github/workflows/ci-cd.yml
@@ -168,12 +168,6 @@ jobs:
           CLOUDFLARE_ACCOUNT_ID: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
         run: if [ "${{ github.ref }}" = "refs/heads/main" ]; then npm run deploy -- --env production; else npm run deploy -- --env staging; fi
 
-      - name: Run Integration Tests
-        working-directory: pages
-        run: |
-          echo "Running integration tests..."
-          npm run test:integration
-
       - name: Run E2E Tests Against Staging
         if: github.event.pull_request.base.ref == 'main'
         working-directory: pages
@@ -185,7 +179,7 @@ jobs:
           echo "Waiting for staging deployment to be ready..."
           sleep 30  # Give CF Pages time to deploy
 
-          # Run E2E tests with extension
+          # Run E2E tests with extension against staging
           echo "Running E2E tests against staging..."
           npm run test:e2e
