+++ pages/package.json
@@ -6,19 +6,15 @@
     "dev": "vite",
     "build": "npm run build:web && npm run build:extensions",
     "build:web": "tsc && vite build",
-    "test:integration": "jest",
+    "test": "jest",
+    "test:coverage": "jest --coverage",
     "test:e2e": "BROWSER=chrome npm run build:extensions && playwright test --trace on",
-    "test:e2e:ui": "BROWSER=chrome npm run build:extensions && playwright test --ui",
-    "test:e2e:debug": "BROWSER=chrome npm run build:extensions && playwright test --debug",
     "validate:manifests": "node scripts/validate-manifests.js",
     "build:extensions": "npm run validate:manifests && BROWSER=${BROWSER:-chrome} MANIFEST_VERSION=${MANIFEST_VERSION:-v3} vite build --mode extension",
     "package:chrome": "cd dist/chrome && rm -f ../../chroniclesync-chrome.zip && zip -r ../../chroniclesync-chrome.zip *",
     "package:firefox": "cd dist/firefox && rm -f ../../chroniclesync-firefox.zip && zip -r ../../chroniclesync-firefox.zip *",
     "package:safari": "bash -e ./scripts/build-safari-extension-direct.sh",
     "preview": "vite preview",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage",
     "lint": "eslint src",
     "lint:fix": "eslint src --fix",
     "deploy": "wrangler pages deploy dist/web --project-name chroniclesync-pages"
